:recycle: Move PlacePolicyScope to its own file.
Why? To keep the size of the tests for the file smaller. Also one file per class is usually a good idea.
