+++ src/api/policies/index.ts
@@ -1,2 +1,3 @@
 export * from './base-policy';
 export * from './place-policy';
+export * from './place-policy-scope';
