+++ src/api/policies/place-policy.ts
@@ -1,68 +1,7 @@
-import { isEmpty, intersection, toInteger } from 'lodash';
-import { Knex } from 'knex';
+import { isEmpty, intersection } from 'lodash';
 
-import { BasePolicy, BasePolicyScope } from '.';
+import { BasePolicy } from '.';
 import { User, Place, UserRoles } from '../models';
-
-export class PlacePolicyScope extends BasePolicyScope {
-	constructor(scope: Knex.QueryBuilder, user: User) {
-		super(scope, user);
-		this.scope = scope
-			.select(['Place.Id'])
-			.leftOuterJoin(
-				'FirstNationAssociation',
-				'Place.Id',
-				'FirstNationAssociation.PlaceId'
-			);
-	}
-
-	resolve() {
-		// without a user passed in, you see nothing
-		if (!this.user) {
-			return this.emptyScope;
-		}
-
-		// Administrators see everything
-		if (this.user.roleList.includes(UserRoles.ADMINISTRATOR)) {
-			return this.scope;
-		}
-
-		// If you don't have one of the site roles, you see nothing
-		if (
-			isEmpty(
-				intersection(this.user.roleList, [
-					UserRoles.SITE_ADMIN,
-					UserRoles.SITE_EDITOR,
-					UserRoles.SITE_VIEWER,
-				])
-			)
-		) {
-			return this.emptyScope;
-		}
-
-		let query = '';
-		if (!isEmpty(this.user.permittedMapSheets)) {
-			const permittedMapSheets = this.user.permittedMapSheets.join("','");
-			query += ` OR NTSMapSheet IN ('${permittedMapSheets}')`;
-		}
-		if (!isEmpty(this.user.permittedCommunityIds)) {
-			const permittedCommunityIds = this.user.permittedCommunityIds.join("','");
-			query += ` OR CommunityId IN (${permittedCommunityIds})`;
-		}
-		if (!isEmpty(this.user.permittedFirstNationsIds)) {
-			const permittedFirstNationsIds =
-				this.user.permittedFirstNationsIds.join("','");
-			query += ` OR [FirstNationAssociation].[FirstNationId] IN (${permittedFirstNationsIds})`;
-		}
-
-		if (!isEmpty(query)) {
-			query += `(${query})`;
-			return this.scope.whereRaw(query);
-		}
-
-		return this.emptyScope;
-	}
-}
 
 export class PlacePolicy extends BasePolicy<Place> {
 	constructor(user: User, record: Place) {
